refactor: #21 dto 클래스 위치 변경
