+++ src/main/java/com/project/kodesalon/model/board/service/BoardService.java
@@ -1,11 +1,11 @@
 package com.project.kodesalon.model.board.service;
 
 import com.project.kodesalon.model.board.domain.Board;
-import com.project.kodesalon.model.board.domain.dto.BoardCreateRequestDto;
 import com.project.kodesalon.model.board.domain.vo.Content;
 import com.project.kodesalon.model.board.domain.vo.Title;
 import com.project.kodesalon.model.board.exception.ForbiddenException;
 import com.project.kodesalon.model.board.repository.BoardRepository;
+import com.project.kodesalon.model.board.service.dto.BoardCreateRequestDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
